test(fix): Fix broken tests
