+++ review_rating/modeling/__main__.py
@@ -3,11 +3,9 @@
 import argparse
 from pathlib import Path
 
-from review_rating.modeling.model_analysis import (
-    evaluate_model,
-    prepare_data,
-    train_model,
-)
+from review_rating.modeling.evaluate import evaluate_model
+from review_rating.modeling.prepare_data import prepare_data
+from review_rating.modeling.train import train_model
 
 DEFAULT_DATA_PATH = "data/a1_RestaurantReviews_HistoricDump.tsv"
