+++ review_rating/tests/conftest.py
@@ -7,19 +7,28 @@ from lib_ml import Preprocessor
 from sklearn.feature_extraction.text import CountVectorizer
 from sklearn.model_selection import train_test_split
 
-from review_rating.modeling import model_analysis
-
-MODEL_PATH = "sentiment_model"
+from review_rating.modeling.train import train_model
+from review_rating.modeling.prepare_data import prepare_data
 
 
 @pytest.fixture(scope="session", autouse=True)
 def build_artifacts():
     """Run training script once per test session to generate model/vectorizer."""
-    model_analysis.run_pipeline(
-        "data/raw/a1_RestaurantReviews_HistoricDump.tsv", MODEL_PATH, 0.2, 42
+    prepare_data(
+        input_path="data/raw/a1_RestaurantReviews_HistoricDump.tsv",
+        output_dir="processed",
+        test_split=0.2,
+        random_state=42,
+        max_features=1420,
     )
-    assert os.path.exists(f"{MODEL_PATH}.pkl"), "Model file not created"
-    assert os.path.exists(f"{MODEL_PATH}_vectorizer.pkl"), "Vectorizer file not created"
+
+    assert os.path.exists(f"processed/processed_data.pk1"), "Data file not created"
+    assert os.path.exists(f"processed/vectorizer.pkl"), "Vectorizer file not created"
+
+    train_model(
+        input_dir="processed",
+        output_model="processed/sentiment_model.pkl",
+    )
 
 
 @pytest.fixture(scope="session")
@@ -61,5 +70,5 @@ def split_data(X_y):
 
 @pytest.fixture(scope="session")
 def classifier(split_data):
-    clf = joblib.load("sentiment_model.pkl")
+    clf = joblib.load("processed/sentiment_model.pkl")
     return clf
