+++ review_rating/tests/test_infrastructure.py
@@ -5,21 +5,21 @@ import joblib
 
 
 def test_model_serialized():
-    assert os.path.exists("sentiment_model.pkl")
-    model = joblib.load("sentiment_model.pkl")
+    assert os.path.exists("processed/sentiment_model.pkl")
+    model = joblib.load("processed/sentiment_model.pkl")
     assert hasattr(model, "predict")
 
 
 def test_vectorizer_serialized():
-    assert os.path.exists("sentiment_model_vectorizer.pkl")
-    vec = pickle.load(open("sentiment_model_vectorizer.pkl", "rb"))
+    assert os.path.exists("processed/vectorizer.pkl")
+    vec = pickle.load(open("processed/vectorizer.pkl", "rb"))
     assert hasattr(vec, "transform")
 
 
 def test_prediction_pipeline(preprocessor):
     text = "This place was not great"
-    vec = pickle.load(open("sentiment_model_vectorizer.pkl", "rb"))
-    model = joblib.load("sentiment_model.pkl")
+    vec = pickle.load(open("processed/vectorizer.pkl", "rb"))
+    model = joblib.load("processed/sentiment_model.pkl")
 
     processed = preprocessor.preprocess(text)
     X = vec.transform([processed]).toarray()
